refactor(documents): move contract description helper to contracts module; add changeset
- Move `buildContractDocumentDescription` from `issue-document.helper.ts` into
  `contracts/helpers/contracts.helper.ts`. The former transitively imports the green-invoice
  provider, which broke the unit test (`@accounter/green-invoice-graphql` entry resolution).
  The helper is pure, so it belongs in the contracts module; its test moves alongside it.
- Apply prettier formatting.
- Add a changeset for the contract-generated document description change.
