+++ packages/server/src/modules/contracts/helpers/__tests__/contracts.helper.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import type { IGetContractsByIdsResult } from '../../../contracts/types.js';
-import { buildContractDocumentDescription } from '../issue-document.helper.js';
+import type { IGetContractsByIdsResult } from '../../types.js';
+import { buildContractDocumentDescription } from '../contracts.helper.js';
 
 function buildContract(overrides: Partial<IGetContractsByIdsResult>): IGetContractsByIdsResult {
   return {
@@ -19,7 +19,7 @@ describe('buildContractDocumentDescription', () => {
   it('uses the billed month for monthly contracts', () => {
     const contract = buildContract({ billing_cycle: 'monthly' });
 
-    // issueMonth is the previous month; the description points at it (issueMonth + 1 → previous month)
+    // The billed month is the issue month itself.
     expect(buildContractDocumentDescription(contract, '2026-05-01')).toBe(
       'GraphQL Hive Pro Plan - May 2026',
     );
